+++ frontend/src/app/(dynamic)/(pages)/account/email-verified/page.tsx
@@ -1,7 +1,7 @@
-import { Route } from '../../../../components/ButtonsAndLinks/Route/Route'
-import { Headline } from '../../../../components/Headline/Headline'
-import { Icon } from '../../../../components/Icon/Icon'
-import { routes } from '../../../../services/routes/routes'
+import { Route } from '../../../../../components/ButtonsAndLinks/Route/Route'
+import { Headline } from '../../../../../components/Headline/Headline'
+import { Icon } from '../../../../../components/Icon/Icon'
+import { routes } from '../../../../../services/routes/routes'
 
 export const metadata = {
 	title: 'Email verified | Siebtes Leben',
