+++ frontend/src/app/(dynamic)/(pages)/layout.tsx
@@ -1,0 +1,22 @@
+import { GlobalFooter } from '../../../components/Navbar/GlobalFooter/GlobalFooter'
+import { Navbar } from '../../../components/Navbar/Navbar/Navbar'
+import { NavbarLogo } from '../../../components/Navbar/NavbarLogo/NavbarLogo'
+
+export const metadata = {
+	title: 'Siebtes Leben',
+	description: '',
+}
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+	return (
+		<>
+			<Navbar>
+				<div className="mr-5">
+					<NavbarLogo />
+				</div>
+			</Navbar>
+			{children}
+			<GlobalFooter />
+		</>
+	)
+}
