+++ frontend/src/app/(dynamic)/(pages)/profile/layout.tsx
@@ -1,8 +1,8 @@
 'use client'
 import { usePathname } from 'next/navigation'
-import { profileLinks } from '../../../../content/profilelinks'
-import { Headline } from '../../../components/Headline/Headline'
-import { ProfileSideBarLink } from '../../../components/Navbar/ProfileSideBarLink/ProfileSideBarLink'
+import { profileLinks } from '../../../../../content/profilelinks'
+import { Headline } from '../../../../components/Headline/Headline'
+import { ProfileSideBarLink } from '../../../../components/Navbar/ProfileSideBarLink/ProfileSideBarLink'
 
 export default function Layout({ children }: { children: React.ReactNode }) {
 	const pathname = usePathname()
