+++ frontend/src/app/(dynamic)/last-will/(pages)/auth/page.tsx
