+++ frontend/src/app/(dynamic)/last-will/(pages)/buy/page.tsx
