+++ frontend/src/app/(dynamic)/last-will/(pages)/layout.tsx
