+++ frontend/src/app/(dynamic)/last-will/editor/heirs/page.tsx
