+++ frontend/src/app/(dynamic)/last-will/editor/inheritance/page.tsx
