+++ frontend/src/app/(dynamic)/last-will/editor/layout.tsx
