+++ frontend/src/app/(dynamic)/last-will/editor/succession/page.tsx
