+++ frontend/src/app/(dynamic)/layout.tsx
