+++ frontend/src/app/(static)/layout.tsx
