+++ frontend/src/app/(static)/misc/faq/[slug]/page.tsx
