+++ frontend/src/app/(static)/misc/faq/page.tsx
