+++ frontend/src/app/(static)/page.tsx
